Fixed missing includes + changed CMakeLists.txt
Changes due to renaming and moving of files
