+++ arch/M68K/M68KInstPrinter.h
@@ -10,9 +10,11 @@
 #include "../../MCRegisterInfo.h"
 #include "../../MCInst.h"
 
+struct SStream;
+
 void M68K_init(MCRegisterInfo *MRI);
 
-void M68K_printInst(MCInst* MI, SStream* O, void* Info);
+void M68K_printInst(MCInst* MI, struct SStream* O, void* Info);
 bool M68K_getInstruction(csh ud, const uint8_t* code, size_t code_len, MCInst* instr, uint16_t* size, uint64_t address, void* info);
 const char* M68K_reg_name(csh handle, unsigned int reg);
 void M68K_get_insn_id(cs_struct* h, cs_insn* insn, unsigned int id);
